requirements.txt
openai>=1.14.0
python-dotenv>=1.0.0
chromadb>=0.4.22
sentence-transformers>=2.6.0
gradio>=4.19.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0
